FRONTEND: Modularización del Sistema Selnic, antes llamado DMS, se esta definiendo la estructura modular de carpetas
