--- frontend/src/components/Header.components.jsx
@@ -1,134 +1,0 @@
-import { Link } from "react-router-dom";
-import useAuth from "../hooks/useAuth.hook";
-
-const Header = () => {
-  const { auth, cerrarSesion } = useAuth();
-
-  return (
-    <header className="py-10 bg-indigo-600">
-      <div className="container mx-auto flex flex-col lg:flex-row justify-between items-center">
-        {/* <h1 className="font-bold text-2xl text-indigo-200 text-center select-none ">
-          Administrador de Documentos en {""}{" "}
-          <span className="text-white font-black">DMS</span>
-        </h1> */}
-        <Link
-          to={"/admin"}
-          className="font-bold text-2xl text-indigo-200 text-center select-none hover:text-white transition-colors duration-300 group"
-        >
-          Administrador de Documentos en{" "}
-          <span className="text-white font-black relative inline-block">
-            DMS
-            <span className="block w-0 group-hover:w-full h-[2px] bg-white transition-all duration-300 absolute bottom-0 left-0"></span>
-          </span>
-        </Link>
-
-        <nav className="flex flex-col items-center lg:flex-row gap-4 mt-5 lg:mt-0 ">
-          <nav className="relative group">
-            <Link
-              // to={"/"}
-              className="text-white text-sm uppercase font-bold  hover:underline"
-            >
-              Trámites
-            </Link>
-
-            <div className="absolute hidden group-hover:block text-white text-sm uppercase font-bold bg-indigo-500 p-2 shadow-lg rounded-lg">
-              <Link
-                to={"/admin/ingresar"}
-                className="block px-4 py-2 hover:bg-indigo-700"
-              >
-                Ingresar Trámites
-              </Link>
-
-              {/* Solo visible para COODINADOR */}
-              {auth.rol === "COORDINADOR" && (
-                <Link
-                  to={"/admin/asignar-reasignar"}
-                  className="block px-4 py-2 hover:bg-indigo-700"
-                >
-                  Asignar/Reasignar Trámites
-                </Link>
-              )}
-
-              {auth.rol !== "DESPACHADOR" && auth.rol !== "COORDINADOR" && (
-                <Link
-                  to={"/admin/asignados"}
-                  className="block px-4 py-2 hover:bg-indigo-700"
-                >
-                  Ver Trámites Asignados
-                </Link>
-              )}
-
-              {/* Solo visible para COORDINADOR */}
-              {auth.rol === "COORDINADOR" && (
-                <Link
-                  to={"/admin/completar-tramite"}
-                  className="block px-4 py-2 hover:bg-indigo-700"
-                >
-                  Trámites por Firmar
-                </Link>
-              )}
-              {/* {auth.rol === "DESPACHADOR" && ( */}
-              {auth.rol !== "REVISOR" && auth.rol !== "USUARIO" && (
-                <Link
-                  to={"/admin/despachar-tramite"}
-                  className="block px-4 py-2 hover:bg-indigo-700"
-                >
-                  Trámites Por Despachar
-                </Link>
-              )}
-              <Link
-                to={"/admin/consultar-tramite"}
-                className="block px-4 py-2 hover:bg-indigo-700"
-              >
-                Consultas Trámites
-              </Link>
-              {/* Solo visible para COORDINADOR y DESPACHADOR*/}
-              {(auth.rol === "COORDINADOR" || auth.rol === "DESPACHADOR") && (
-                <Link
-                  to={"/admin/admin-dms"}
-                  className="block px-4 py-2 hover:bg-indigo-700"
-                >
-                  Configuración
-                </Link>
-              )}
-            </div>
-          </nav>
-
-          <nav className="relative group">
-            <Link
-              // to={"/admin/perfil"}
-              className="text-white text-sm uppercase font-bold hover:underline"
-            >
-              Perfil
-            </Link>
-
-            <div className="absolute hidden group-hover:block text-white text-sm uppercase font-bold bg-indigo-500 p-2 shadow-lg rounded-lg">
-              <Link
-                to={"/admin/perfil"}
-                className="block px-4 py-2 hover:bg-indigo-700"
-              >
-                Editar Pérfil
-              </Link>
-              <Link
-                to={"/admin/cambiar-password"}
-                className="block px-4 py-2 hover:bg-indigo-700"
-              >
-                Cambiar Password
-              </Link>
-            </div>
-          </nav>
-
-          <button
-            type="button"
-            className="text-white text-sm uppercase font-bold hover:underline"
-            onClick={cerrarSesion}
-          >
-            Cerrar Sesión
-          </button>
-        </nav>
-      </div>
-    </header>
-  );
-};
-
-export default Header;
