+++ frontend/src/modules/document-management/pages/AdministrarTramites.pages.jsx
@@ -1,0 +1,33 @@
+import { useState } from "react";
+import Formulario from "../../../components/Formulario.components";
+import ListadoTramites from "../../../components/ListadoTramites.components";
+
+const AdministrarTramites = () => {
+  const [mostrarFormulario, setMostrarFormulario] = useState(false);
+
+  return (
+    <div className="flex flex-col md:flex-row">
+      <button
+        type="button"
+        className="bg-indigo-600 text-white font-bold uppercase mx-10 p-3 rounded-md mb-10 md:hidden"
+        onClick={() => {
+          setMostrarFormulario(!mostrarFormulario); // Para quecada vez que se presione haga lo contrario de mostrar formulario
+        }}
+      >
+        {mostrarFormulario ? "Ocultar Formulario" : "Mostrar Formulario"}
+      </button>
+      <div
+        className={`${
+          mostrarFormulario ? "block" : "hidden"
+        } md:block md:w-1/2 lg:w-2/5 `}
+      >
+        <Formulario />
+      </div>
+      <div className="md:w-1/2 lg:w-3/5">
+        <ListadoTramites />
+      </div>
+    </div>
+  );
+};
+
+export default AdministrarTramites;
